- In Dispatcher, avoid recalculate the information if it has not changed. - An update bandwidth limit of 0 means no limit. - Change ZAFunction fields from double to float, to reduce message size about 40%-50%. - Use std::array instead of C++ array. - In FSPDispatcher, reworked forwarding algorithm to avoid sending more than one task to each node, where posible. - Bug fix: MMP scheduler did not set the maximum queue end time in the generated information. - Use Boost::Date_time in PerformanceStatistics instead of std::clock() to improve precision.

+++ include/ZAFunction.hpp
@@ -73,9 +73,9 @@ public:
 
         MSGPACK_DEFINE(leftEndpoint, x, y, z1, z2);
 
-        double leftEndpoint;
+        float leftEndpoint;
         // z1 is the sum of the independent term in L = x/a + z1, while z2 is the independent part in the other functions
-        double x, y, z1, z2;   // L = x/a + y*a + z1 + z2
+        float x, y, z1, z2;   // L = x/a + y*a + z1 + z2
     };
 
     typedef std::vector<SubFunction> PieceVector;
